Moving page code to TradesWidget
